# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromptSnapshotResponse": {
        "additionalProperties": false,
        "properties": {
          "activated_skills": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Activated Skills",
            "type": "array"
          },
          "agents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Agents",
            "type": "array"
          },
          "prompt_revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Revision"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "tools": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "prompt_revision",
          "agents",
          "tools",
          "activated_skills"
        ],
        "title": "PromptSnapshotResponse",
        "type": "object"
      },
...
      "RunResponse": {
        "additionalProperties": false,
        "properties": {
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "cancel_requested": {
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "epochs": {
            "items": {
              "$ref": "#/components/schemas/RunEpochResponse"
            },
            "title": "Epochs",
            "type": "array"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/RunEventResponse"
            },
            "title": "Events",
            "type": "array"
          },
          "final_output": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Final Output"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "goal_state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal State"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input": {
            "title": "Input"
          },
          "interruptions": {
            "items": {
              "$ref": "#/components/schemas/RunInterruptionResponse"
            },
            "title": "Interruptions",
            "type": "array"
          },
          "items": {
            "items": {
              "discriminator": {
                "mapping": {
                  "handoff_call_item": "#/components/schemas/HandoffCallRunItem",
                  "handoff_output_item": "#/components/schemas/HandoffOutputRunItem",
                  "mcp_approval_request_item": "#/components/schemas/McpApprovalRequestRunItem",
                  "mcp_approval_response_item": "#/components/schemas/McpApprovalResponseRunItem",
                  "mcp_list_tools_item": "#/components/schemas/McpListToolsRunItem",
                  "message_output_item": "#/components/schemas/MessageRunItem",
                  "reasoning_item": "#/components/schemas/ReasoningRunItem",
                  "tool_approval_item": "#/components/schemas/ToolApprovalRunItem",
                  "tool_call_item": "#/components/schemas/ToolCallRunItem",
                  "tool_call_output_item": "#/components/schemas/ToolOutputRunItem",
                  "tool_search_call_item": "#/components/schemas/ToolSearchCallRunItem",
                  "tool_search_output_item": "#/components/schemas/ToolSearchOutputRunItem"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MessageRunItem"
                },
                {
                  "$ref": "#/components/schemas/ToolCallRunItem"
                },
                {
                  "$ref": "#/components/schemas/ToolOutputRunItem"
                },
                {
                  "$ref": "#/components/schemas/HandoffCallRunItem"
                },
                {
                  "$ref": "#/components/schemas/HandoffOutputRunItem"
                },
                {
                  "$ref": "#/components/schemas/ReasoningRunItem"
                },
                {
                  "$ref": "#/components/schemas/ToolApprovalRunItem"
                },
                {
                  "$ref": "#/components/schemas/ToolSearchCallRunItem"
                },
                {
                  "$ref": "#/components/schemas/ToolSearchOutputRunItem"
                },
                {
                  "$ref": "#/components/schemas/McpListToolsRunItem"
                },
                {
                  "$ref": "#/components/schemas/McpApprovalRequestRunItem"
                },
                {
                  "$ref": "#/components/schemas/McpApprovalResponseRunItem"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "last_agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Agent Name"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "paused",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "tool_attempts": {
            "items": {
              "$ref": "#/components/schemas/ToolAttemptResponse"
            },
            "title": "Tool Attempts",
            "type": "array"
          },
          "usage": {
            "additionalProperties": true,
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "id",
          "conversation_id",
          "agent_name",
          "status",
          "input",
          "final_output",
          "last_agent_name",
          "usage",
          "error",
          "cancel_requested",
          "created_at",
          "started_at",
          "finished_at",
          "items",
          "events",
          "interruptions",
          "epochs",
          "tool_attempts",
          "goal_state"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "SteeringMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 100000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "title": "SteeringMessageRequest",
        "type": "object"
      },
      "SteeringMessageResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "status"
        ],
        "title": "SteeringMessageResponse",
        "type": "object"
      },
      "StopAndAnswerResponse": {
        "additionalProperties": false,
        "properties": {
          "answer_run": {
            "$ref": "#/components/schemas/RunResponse"
          },
          "stopped_run": {
            "$ref": "#/components/schemas/RunResponse"
          }
        },
        "required": [
          "stopped_run",
          "answer_run"
        ],
        "title": "StopAndAnswerResponse",
        "type": "object"
      },
...
      "WorkspaceFileContentResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "modified_at": {
            "format": "date-time",
            "title": "Modified At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Id"
          },
          "note_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Name"
          },
          "paper_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Id"
          },
          "paper_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Name"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "path",
          "name",
          "media_type",
          "size_bytes",
          "modified_at",
          "tags",
          "kind",
          "content"
        ],
        "title": "WorkspaceFileContentResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ScholarWeave API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "conversation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunResponse"
                  },
                  "title": "Response List Runs Api Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "delete": {
        "operationId": "delete_run_api_runs__run_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Run",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": -1,
              "minimum": -1,
              "title": "After",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Events",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/prompt-snapshot": {
      "get": {
        "operationId": "get_prompt_snapshot_api_runs__run_id__prompt_snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Prompt Snapshot",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/steering": {
      "post": {
        "operationId": "steer_run_api_runs__run_id__steering_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SteeringMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SteeringMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Steer Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/stop-and-answer": {
      "post": {
        "operationId": "stop_and_answer_run_api_runs__run_id__stop_and_answer_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StopAndAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop And Answer Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/workspace/files/content": {
      "delete": {
        "operationId": "delete_file_api_workspace_files_content_delete",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete File",
        "tags": [
          "workspace"
        ]
      },
      "get": {
        "operationId": "read_file_api_workspace_files_content_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceFileContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read File",
        "tags": [
          "workspace"
        ]
      },
      "put": {
        "operationId": "write_file_api_workspace_files_content_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceFileWriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceFileContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write File",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/workspace/files/folder": {
      "delete": {
        "operationId": "delete_folder_api_workspace_files_folder_delete",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Folder",
        "tags": [
          "workspace"
        ]
      }
    },
    "/api/workspace/files/notes": {
      "post": {
        "operationId": "create_note_api_workspace_files_notes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceNoteCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceFileContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Note",
        "tags": [
          "workspace"
        ]
      }
    }
  }
}